refactor labels
